add localeSelect component
https://github.com/magickasoft/fit-paradise/issues/63
