+++ messages/ru.json
@@ -7,9 +7,12 @@
       "label": "Кнопка primary"
     }
   },
-  "localeSwitcher": {
+  "localeSelect": {
     "label": "Выберите язык",
-    "locale": "{locale, select, ru {Russia} en {English} other {Unknown}}"
+    "locale": {
+      "ru": "Русский",
+      "en": "Английский"
+    }
   },
   "manifest": {
     "name": "next-intl example"
